Stop scaling stroke-width on non-graphics elements
